refactor(conf): move STREAM_CHAINS to the neutral aleph.vm.chains
conf.py is neutral ground shared by the agent and the supervisor, yet it
imported STREAM_CHAINS from aleph.vm.agent.chain. Move the whole chain
module (ChainInfo, STREAM_CHAINS, get_chain) to aleph.vm.chains, which
only depends on aleph_message and pydantic, so conf never imports
agent.* again. Flagged as a Phase A follow-up.
